pyproject.toml
[project]
name = "ai-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.7",
    "docling>=2.31.0",
    "logfire>=3.14.1",
    "loguru>=0.7.3",
    "mlx-vlm>=0.0.14",
    "openai>=1.76.0",
    "prompt-toolkit>=3.0.51",
    "pydantic-ai-slim[openai]>=0.1.8",
    "pydantic-settings>=2.9.1",
    "python-telegram-bot>=22.0",
    "tiktoken>=0.9.0",
    "typer>=0.15.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.400",
    "ruff>=0.11.7",
]
